# perf: pass transpiled module to deno_core as known string