When using transpile cache:

```
% hyperfine --warmup 4 "../deno/target/release/deno run -A main.ts" "../deno/target/release/deno_main run -A main.ts"
Benchmark 1: ../deno/target/release/deno run -A main.ts
  Time (mean ± σ):     101.8 ms ±   1.0 ms    [User: 82.7 ms, System: 24.8 ms]
  Range (min … max):    99.6 ms … 103.9 ms    28 runs
 
Benchmark 2: ../deno/target/release/deno_main run -A main.ts
  Time (mean ± σ):     107.4 ms ±   1.6 ms    [User: 86.9 ms, System: 25.9 ms]
  Range (min … max):   104.3 ms … 112.1 ms    27 runs
 
Summary
  ../deno/target/release/deno run -A main.ts ran
    1.05 ± 0.02 times faster than ../deno/target/release/deno_main run -A main.ts
```

No transpile cache:

```
% DENO_DIR=$pwd/deno_dir hyperfine --prepare "rm -rf deno_dir" --warmup 4 "../deno/target/release/deno run -A main.ts" "../deno/target/release/deno_main run -A main.ts"
Benchmark 1: ../deno/target/release/deno run -A main.ts
  Time (mean ± σ):      2.479 s ±  0.016 s    [User: 4.259 s, System: 0.132 s]
  Range (min … max):    2.458 s …  2.508 s    10 runs
 
Benchmark 2: ../deno/target/release/deno_main run -A main.ts
  Time (mean ± σ):      2.493 s ±  0.012 s    [User: 4.273 s, System: 0.134 s]
  Range (min … max):    2.478 s …  2.514 s    10 runs
 
Summary
  ../deno/target/release/deno run -A main.ts ran
    1.01 ± 0.01 times faster than ../deno/target/release/deno_main run -A main.ts
```

I noticed this when working on the deno compile changes because I had a regression in performance when just providing deno_core with bytes.

<details>
<summary>Files</summary>

typescript1.ts and typescript2.ts are two copies of the typescript compiler as typescript files (not JS files)

```ts
import { ts as ts1 } from "./typescript1.ts";
import { ts as ts2 } from "./typescript2.ts";

console.log(ts1.version);
console.log(ts2.version);

```
</details>